Fixes for Terraform Repositories
Why these changes are being introduced:
Renovate/Mend has some issues with the pre-commit hooks dependencies in
our Terraform repositories (see PR#15 in mitlib-tf-workloads-tacos as
an example). This is an attempt to address that cycling behavior.

How this addresses that need:
* Add additional settings to the pre-commit configuration to see if
this will fix the cycling behavior.

Side effects of this change:
None.